+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 0.24.0
 
+* `Common`:
+  * Rename `Progress` to more common `Percentage`. `Progress` now is typealias
+  * Fix of `Progress.compareTo` extension
+
 ## 0.23.2
 
 * `Versions`: